Support spelling-private-dict-file without spelling-dict
Allow Pylint to be configured with a private dictionary (wordlist) file
without specifying a language dictionary to load.  The primary use case
for this configuration is to facilitate Pylint testing without an
implicit dependency on the installed enchant backends (aspell, ispell,
myspell, etc.) installed wordlists, or locale.  It may also be useful
for projects which choose to use a specific wordlist for consistency.

Note:  This is a behavior change for projects which configure
spelling-private-dict-file but do not expect it to be used.  This can be
fixed by commenting out the undesired spelling-private-dict-file. I contend
that the benefits of the new behavior (for testability, usefulness, and
being closer to expected behavior) outweigh the potential breakage.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>