Install pyenchant to test spelling checker
Most of the spelling checker code is not tested when pyenchant is not
installed.  To improve the test coverage, install pyenchant.

Note:  This change requires users running tox (e.g. via `make tests`) on
systems without a pyenchant wheel (e.g. Linux, BSD) to ensure the
enchant library is installed on their system.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>